recipy_website/requirements.txt
Django>=4.0.0, <5.0.0
python-dotenv
django_elasticsearch_dsl
django-mailgun-mime
